+++ changelog.md
@@ -1,5 +1,13 @@
 # Backpack changelog
 
+## UNRELEASED
+
+**Added:**
+- bpk-component-button:
+- bpk-mixins:
+- bpk-tokens:
+  - Added `destructive` button style
+
 ## 2016-11-09 - Docked inputs & selects 
 
 **Changed:**